atexit + _signal: extract into module/atexit/ + module/_signal/
importing.rs shrinks by 532 lines.
