+++ pyre/pyre-interpreter/src/module/_signal/moduledef.rs
@@ -1,0 +1,7 @@
+//! _signal module definition — PyPy: pypy/module/signal/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_signal::register_module(ns);
+}
